# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Project": {
        "properties": {
          "created": {
            "default": "",
            "title": "Created",
            "type": "string"
          },
          "grid_config": {
            "default": "",
            "title": "Grid Config",
            "type": "string"
          },
          "id": {
            "default": "",
            "title": "Id",
            "type": "string"
          },
          "ligand_source": {
            "default": "",
            "title": "Ligand Source",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "page_inputs": {
            "additionalProperties": true,
            "default": {},
            "title": "Page Inputs",
            "type": "object"
          },
          "pdb_id": {
            "default": "",
            "title": "Pdb Id",
            "type": "string"
          },
          "pipeline_state": {
            "additionalProperties": true,
            "default": {},
            "title": "Pipeline State",
            "type": "object"
          },
          "receptor_path": {
            "default": "",
            "title": "Receptor Path",
            "type": "string"
          },
          "session_data": {
            "additionalProperties": true,
            "default": {},
            "title": "Session Data",
            "type": "object"
          },
          "target_name": {
            "default": "",
            "title": "Target Name",
            "type": "string"
          },
          "updated": {
            "default": "",
            "title": "Updated",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Project",
        "type": "object"
      },
      "ReportRequest": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "custom_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-form notes to include",
            "title": "Custom Text"
          },
          "format": {
            "default": "both",
            "description": "'pdf' | 'html' | 'both'",
            "title": "Format",
            "type": "string"
          },
          "output_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Directory to save reports",
            "title": "Output Dir"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "results_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Auto-collect from docking results dir",
            "title": "Results Dir"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Custom sections to include",
            "title": "Sections"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Report title",
            "title": "Title"
          }
        },
        "title": "ReportRequest",
        "type": "object"
      },
      "ReportResponse": {
        "properties": {
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "paths": {
            "description": "Generated report file paths",
            "items": {
              "type": "string"
            },
            "title": "Paths",
            "type": "array"
          }
        },
        "title": "ReportResponse",
        "type": "object"
      },
      "ResolveRequest": {
        "properties": {
          "input": {
            "description": "SMILES, InChI, compound name, CAS number, or MOL block",
            "title": "Input",
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
      "ResolveResponse": {
        "properties": {
          "input_type": {
            "title": "Input Type",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "smiles": {
            "title": "Smiles",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "smiles",
          "input_type"
        ],
        "title": "ResolveResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Drug Discovery Suite — REST + WebSocket API",
    "title": "Helix Core Backend",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/delete/{project_id}": {
      "delete": {
        "description": "Delete a project by ID.",
        "operationId": "delete_project_api_projects_delete__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "Project Management"
        ]
      }
    },
...
    "/api/projects/load/{project_id}": {
      "get": {
        "description": "Load a single project by ID.",
        "operationId": "load_project_api_projects_load__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Project",
        "tags": [
          "Project Management"
        ]
      }
    },
    "/api/projects/save": {
      "post": {
        "description": "Create or overwrite a named project.",
        "operationId": "save_project_api_projects_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Project",
        "tags": [
          "Project Management"
        ]
      }
    },
    "/api/report/download": {
      "get": {
        "description": "Download a generated report file.",
        "operationId": "download_report_api_report_download_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Report",
        "tags": [
          "Report Generation"
        ]
      }
    },
    "/api/report/generate": {
      "post": {
        "description": "Generate a PDF or HTML report from screening results.",
        "operationId": "generate_report_api_report_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Report",
        "tags": [
          "Report Generation"
        ]
      }
    },
    "/api/resolve/batch": {
      "post": {
        "description": "Resolve multiple molecular identifiers in one call.\n\nRequest body: { \"inputs\": [\"aspirin\", \"InChI=1S/...\", \"CC(=O)O\", \"50-78-2\"] }\nReturns: { \"results\": [...], \"resolved\": N, \"failed\": N }",
        "operationId": "resolve_batch_api_resolve_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Batch Api Resolve Batch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Batch",
        "tags": [
          "Molecule Resolver"
        ]
      }
    },
    "/api/resolve/molecule": {
      "post": {
        "description": "Resolve a molecular identifier (SMILES, InChI, name, CAS, MOL block) to canonical SMILES.",
        "operationId": "resolve_molecule_api_resolve_molecule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Molecule",
        "tags": [
          "Molecule Resolver"
        ]
      }
    },
...
  }
}